# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "地端隱私多模態 RAG 專題 API",
    "description": "以多模態 RAG 打造的畫作導覽助理，以及工廠機械加工圖助理（Ortho2CAD 三視圖→3D、庫存 Text-to-SQL、Timefold 生產排程）。SSE 事件見 shared/sse_events.md。",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/images": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Upload Image",
        "operationId": "upload_image_api_v1_images_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_v1_images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/{image_id}": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Get Image",
        "operationId": "get_image_api_v1_images__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          }
        }
      }
    },
    "/api/v1/search/image": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search Image",
        "operationId": "search_image_api_v1_search_image_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageSearchResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/text": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search Text",
        "operationId": "search_text_api_v1_search_text_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Q"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextSearchResponse"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          }
        }
      }
    },
    "/api/v1/artworks": {
      "get": {
        "tags": [
          "artworks"
        ],
        "summary": "List Artworks",
        "operationId": "list_artworks_api_v1_artworks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkListResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artworks/{artwork_id}": {
      "get": {
        "tags": [
          "artworks"
        ],
        "summary": "Get Artwork",
        "operationId": "get_artwork_api_v1_artworks__artwork_id__get",
        "parameters": [
          {
            "name": "artwork_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artwork Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkDetail"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          }
        }
      }
    },
    "/api/v1/artworks/{artwork_id}/image": {
      "get": {
        "tags": [
          "artworks"
        ],
        "summary": "Get Artwork Image",
        "operationId": "get_artwork_image_api_v1_artworks__artwork_id__image_get",
        "parameters": [
          {
            "name": "artwork_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artwork Id"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "full",
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          }
        }
      }
    },
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE：sources／token／done／error",
            "content": {
              "text/event-stream": {}
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parts": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "List Parts",
        "operationId": "list_parts_api_v1_parts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartListResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parts/{part_id}": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "Get Part",
        "operationId": "get_part_api_v1_parts__part_id__get",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Part Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartDetail"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          }
        }
      }
    },
    "/api/v1/parts/{part_id}/drawing": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "Get Part Drawing",
        "operationId": "get_part_drawing_api_v1_parts__part_id__drawing_get",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Part Id"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "full",
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          }
        }
      }
    },
    "/api/v1/parts/{part_id}/model.{ext}": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "Get Part Model",
        "description": "知識庫零件的標準 3D 模型（由 kb/cad/<id>.py 產生）：stl 給前端 3D 檢視、step 給 CAD 軟體。",
        "operationId": "get_part_model_api_v1_parts__part_id__model__ext__get",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Part Id"
            }
          },
          {
            "name": "ext",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ext"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          }
        }
      }
    },
    "/api/v1/search/drawing": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search Drawing",
        "operationId": "search_drawing_api_v1_search_drawing_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrawingSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrawingSearchResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_image_api_v1_images_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_v1_images_post"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "ImageSearchRequest": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          }
        },
        "type": "object",
        "required": [
          "image_id"
        ],
        "title": "ImageSearchRequest"
      },
...
      "ImageUploadResponse": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "width",
          "height"
        ],
        "title": "ImageUploadResponse"
      },
...
    }
  }
}